refactor(kotlin): mirror rust src layout (client/protocol/crypto)
Split the flat com.solana.mpp package into com.solana.mpp.client (Charge,
HttpClient), com.solana.mpp.protocol (Headers, Models, CanonicalJson),
and com.solana.mpp.crypto (Base58, Base64Url, Ed25519, SolanaSigner,
Transaction, Instructions). Test sources mirror the layout. Update the
tests/interop/kotlin-client adapter imports to match. Add examples/
with a minimal ChargeClient as the starting point for the M2 Solana
Seeker demo. gradle check (incl. 90% Jacoco gate) passes.

Refs lgalabru review on PR #105.
