+++ kotlin/src/main/kotlin/com/solana/mpp/crypto/Base58.kt
@@ -1,4 +1,6 @@
-package com.solana.mpp
+package com.solana.mpp.crypto
+
+import com.solana.mpp.protocol.MppException
 
 import java.math.BigInteger
