+++ kotlin/src/main/kotlin/com/solana/mpp/crypto/SolanaSigner.kt
@@ -1,4 +1,6 @@
-package com.solana.mpp
+package com.solana.mpp.crypto
+
+import com.solana.mpp.protocol.MppException
 
 import java.security.KeyPair
