+++ kotlin/src/main/kotlin/com/solana/mpp/protocol/Headers.kt
@@ -1,4 +1,6 @@
-package com.solana.mpp
+package com.solana.mpp.protocol
+
+import com.solana.mpp.crypto.*
 
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.encodeToJsonElement
