+++ kotlin/src/test/kotlin/com/solana/mpp/protocol/HeadersTest.kt
@@ -1,4 +1,6 @@
-package com.solana.mpp
+package com.solana.mpp.protocol
+
+import com.solana.mpp.crypto.*
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
